Remove duplicate drops
Some of these duplicate drops should be kept,
but are removed for the sake of efficiency.

For example, Stumpy (3220000) should drop multiple Tree Branch (4000003) and Leaf (4000005), but the duplicate have been removed.
To fix this, one has to go through the removed drops in this commit and manually reintroduce the drops that should have been kept. The "RemoveDuplicateDrops" tool simplifies this a little bit as it produces a file with all the removed drops. As of writing this, it removed duplicate drops from 415 mobs, for a total of 7331 removed drops.

The following SQL query finds the amount of duplicate drops in the database:
SELECT COUNT(*) total_mobs_with_duplicate_items, SUM(extra_item_drops) AS total_duplicate_items
FROM (SELECT dropperid, SUM(extras) AS extra_item_drops
	FROM
		(SELECT dropperid, itemid, COUNT(*) - 1 AS extras
		FROM lb_drop_data
		GROUP BY dropperid, itemid
		HAVING COUNT(*) > 1) AS i
		GROUP BY dropperid
		ORDER BY dropperid
) AS i2